[refractor]UrlUtil
