--- src/main/java/com/github/jsoncat/common/util/UrlUtil.java
@@ -1,69 +1,0 @@
-package com.github.jsoncat.common.util;
-
-import io.netty.handler.codec.http.QueryStringDecoder;
-import org.apache.commons.codec.CharEncoding;
-import org.apache.commons.codec.Charsets;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- * @author shuang.kou
- * @createTime 2020年09月25日 16:00:00
- **/
-public class UrlUtil {
-
-
-    /**
-     * get the parameters of uri
-     */
-    public static Map<String, String> getQueryParams(String uri) {
-        QueryStringDecoder queryDecoder = new QueryStringDecoder(uri, Charsets.toCharset(CharEncoding.UTF_8));
-        Map<String, List<String>> parameters = queryDecoder.parameters();
-        Map<String, String> queryParams = new HashMap<>();
-        for (Map.Entry<String, List<String>> attr : parameters.entrySet()) {
-            for (String attrVal : attr.getValue()) {
-                queryParams.put(attr.getKey(), attrVal);
-            }
-        }
-        return queryParams;
-    }
-
-    /**
-     * get the decoded path of uri
-     */
-    public static String getRequestPath(String uri) {
-        QueryStringDecoder queryDecoder = new QueryStringDecoder(uri, Charsets.toCharset(CharEncoding.UTF_8));
-        return queryDecoder.path();
-    }
-
-
-    /**
-     * format the url
-     * for example : "/user/{name}" -> "^/user/[\u4e00-\u9fa5_a-zA-Z0-9]+/?$"
-     */
-    public static String formatUrl(String url) {
-        // replace {xxx} placeholders with regular expressions matching Chinese, English letters and numbers, and underscores
-        String originPattern = url.replaceAll("(\\{\\w+})", "[\\\\u4e00-\\\\u9fa5_a-zA-Z0-9]+");
-        String pattern = "^" + originPattern + "/?$";
-        return pattern.replaceAll("/+", "/");
-    }
-
-    /**
-     * Match the request path parameter to the URL parameter
-     *
-     * eg: requestPath="/user/1" url="/user/{id}"
-     * this method will return:{"id" -> "1","user" -> "user"}
-     */
-    public static Map<String, String> getUrlParameterMappings(String requestPath, String url) {
-        String[] requestParams = requestPath.split("/");
-        String[] urlParams = url.split("/");
-        Map<String, String> urlParameterMappings = new HashMap<>();
-        for (int i = 1; i < urlParams.length; i++) {
-            urlParameterMappings.put(urlParams[i].replace("{", "").replace("}", ""), requestParams[i]);
-        }
-        return urlParameterMappings;
-    }
-
-}
